Issue #4: chdir when executing before.sh/after.sh
This is re-factoring.

Use Enum instead of array for STATES.
Enum is well-known design pattern to define a state machine.
Previously we couldn't use it, because we were Python-2 app, and Python-2 didn't support Enum.
But now nothing prevents us from using Enum, when we are Python-3 app.

I need to push this re-factoring prior to actual functional change for Issue #4,
because I'm going to add new states to state machine.